// src/app/definitions/interfaces/order-options.ts
/*
 *  Copyright 2019 Nalej
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
 *  You may obtain a copy of the License at
 *      http://www.apache.org/licenses/LICENSE-2.0
 *  Unless required by applicable law or agreed to in writing, software
 *  distributed under the License is distributed on an "AS IS" BASIS,
 *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 */
import { Order } from '../enums/order.enum';
/**
 *  OrderOptions is used to indicate by which field it is going to be ordered and if the order is asc or des
 */
export interface OrderOptions {
  /**
   * To indicate by which field it is going to be ordered
   */
  field?: string;
  /**
   * Order is used to define the ordering of the results
   */
  order?: Order;
}